Corrected Callback and implemented File Lock logic
Fixes #39
It first checks for a filename with the orderId if not exists proceed to create a file and continue invoice process.

After all logic is executed it removes the file. Cause of this $order->exists() $takes care of it next time.

A new Issue is introduced as when redirect to confirm page with no orderId it doesnt come to where we want.